+++ src/components/pages/PostProject/PostInfo.tsx
@@ -1,22 +1,12 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
+import InputNumber from './InputNumber';
 
 type Props = {};
 
 const PostInfo = (props: Props) => {
-  const [value, setValue] = useState(0);
   const careerList = ['Backend', 'Frontend', 'Designer', 'PM'];
-  const handleIncrement = () => {
-    if (value !== 10) {
-      setValue(value + 1);
-    }
-  };
 
-  const handleDecrement = () => {
-    if (value !== 0) {
-      setValue(value - 1);
-    }
-  };
   return (
     <ProjectInfoContainer>
       <Row>
@@ -32,24 +22,7 @@ const PostInfo = (props: Props) => {
       <h3>모집 인원</h3>
       <Row>
         {careerList.map((career: any) => {
-          return (
-            <ProjectNumberContainer key={career}>
-              <ProjectInfoLabel htmlFor={`${career}Number`}>{career}</ProjectInfoLabel>
-              <ProjectNumber>
-                <PlusNumberBtn onClick={handleDecrement}>-</PlusNumberBtn>
-                <ProjectInfoInput
-                  type="number"
-                  defaultValue={value}
-                  min={0}
-                  max={10}
-                  step={1}
-                  placeholder="0"
-                  id={`${career}Number`}
-                />
-                <MinusNumberBtn onClick={handleIncrement}>+</MinusNumberBtn>
-              </ProjectNumber>
-            </ProjectNumberContainer>
-          );
+          return <InputNumber key={career} career={career} />;
         })}
       </Row>
     </ProjectInfoContainer>
@@ -69,38 +42,10 @@ const Row = styled.div`
   flex-wrap: wrap;
   gap: 20px;
 `;
-const ProjectNumberContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  gap: 10px;
-`;
-const ProjectNumber = styled.div`
-  position: relative;
-  input::-webkit-inner-spin-button,
-  input::-webkit-outer-spin-button {
-    opacity: 0;
-  }
-`;
 const ProjectInfoInput = styled.input`
   padding: 0 15px;
   text-align: center;
 `;
 const ProjectInfoLabel = styled.label`
   margin-right: 10px;
-`;
-const PlusNumberBtn = styled.button`
-  width: 15px;
-  text-align: center;
-  position: absolute;
-  left: 0;
-  top: 50%;
-  transform: translateY(-50%);
-`;
-const MinusNumberBtn = styled.button`
-  width: 15px;
-  text-align: center;
-  position: absolute;
-  right: 0;
-  top: 50%;
-  transform: translateY(-50%);
 `;
